Database authentication (#14)
* Added work for token authentication

Current work on new system for getting OAuth Tokens. Only need a java runner not compiler as they are compiled in .jar already (token-dispenser.jar)

* Changed gplaycli to work with new token  auth

Also changed python_static_analyzer to be a submodule

* Left debug code in

oops

* Refactor

small fixes

* Added authentication for Database and added description text to metadata

Noticed we were just allowing anyone access to the database, so changed to contain a root user. The user/pass is in constants, right now its just filler but will probably want to obfuscate in future somehow.

* Small debug

return_dataframe function was not working with scraper. Should be now.