Semi-Stable Changes - Decompiler compression, Controller options
Added more options to controller to do a mix of download, decompile, and analyze options. Still need to change down_and_analyze mode to also decompile. Future, work in some sort of check to see if an app should be decompiled or not. Currently works in a few tests, not comprehensive yet. Added the keeping of only smali files and compression. Can change to include more file types in the future throguh changing "rm.sh". Right now, controller will download and then decompile, delete non-smali files, and compress all in the apps_decompiled directory. Does not lend well to parallelization right now becuase of all of the os calls and bash commands involved.

Also added the file that does analysis of the file sizes for determining best way forward with pre-decompiling apps.